chore(refactor): split schedule history component into two, split tests as well
