+++ src/curated-corpus/components/ScheduleHistory/ScheduleHistory.tsx
@@ -1,13 +1,10 @@
 import React, { useState } from 'react';
-import { Grid, Typography, Collapse } from '@material-ui/core';
-import { DateTime } from 'luxon';
+import { Grid, Collapse } from '@material-ui/core';
 import { ApprovedCorpusItemScheduledSurfaceHistory } from '../../../api/generatedTypes';
 import { useStyles } from './ScheduleHistory.styles';
-import {
-  getCuratorNameFromLdap,
-  getScheduledSurfaceName,
-} from '../../helpers/helperFunctions';
+
 import { Button } from '../../../_shared/components';
+import { ScheduleHistoryEntries } from '../ScheduleHistoryEntries/ScheduleHistoryEntries';
 
 interface ScheduleHistory {
   /**
@@ -32,12 +29,6 @@ export const ScheduleHistory: React.FC<ScheduleHistory> = (
     ? 'Hide recent scheduled runs'
     : 'View recent scheduled runs';
 
-  const getDisplayDate = (date: string) => {
-    return DateTime.fromFormat(date, 'yyyy-MM-dd')
-      .setLocale('en')
-      .toLocaleString(DateTime.DATE_FULL);
-  };
-
   return (
     <Grid container>
       <Button
@@ -57,34 +48,7 @@ export const ScheduleHistory: React.FC<ScheduleHistory> = (
         unmountOnExit
         className={classes.collapse}
       >
-        {data.map((item: ApprovedCorpusItemScheduledSurfaceHistory) => {
-          return (
-            <Grid
-              key={item.externalId}
-              item
-              xs={12}
-              className={classes.scheduledHistoryWrapper}
-            >
-              <Grid container justifyContent="space-between">
-                <Grid item xs={4}>
-                  <Typography variant="body2">
-                    {getScheduledSurfaceName(item.scheduledSurfaceGuid)}
-                  </Typography>
-                </Grid>
-                <Grid item xs={4}>
-                  <Typography variant="body2">
-                    {getCuratorNameFromLdap(item.createdBy)}
-                  </Typography>
-                </Grid>
-                <Grid item xs={4}>
-                  <Typography variant="body2">
-                    {getDisplayDate(item.scheduledDate)}
-                  </Typography>
-                </Grid>
-              </Grid>
-            </Grid>
-          );
-        })}
+        <ScheduleHistoryEntries data={data} />
       </Collapse>
     </Grid>
   );
